pyproject.toml
[tool.poetry]
name = "paque"
version = "0.1.1"
description = "paque simplifies running simple workflows you want to run. It offers a few features of `make`, but removing most of its power"
authors = ["Ruben Berenguel <ruben+poetry@mostlymaths.net>"]
license = "MIT"
readme = "README.md"

repository = "https://github.com/rberenguel/paque"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3"
cleo = "^0.8.0"
colorlog = "^4.1.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
flake8 = "^3.7.9"
isort = "^4.3.21"
black = "^19.10b0"
mypy = "^0.770"
monkeytype = "^19.11.2"

[tool.poetry.scripts]
paque = 'paque.paque:paque'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

